Makes `env.js` work.
Since only HTTP works for sync calls and everything in here is a sync call, I removed the code that provided other connections as ethrpc will throw an error if you try to make a sync call with IPC or WS (connect in particular).  We could change this script to support async, but for this initial implementation I chose to keep it simple and explicitly only support sync (HTTP).