Fixes some tests.
Also adds an assertion in the code.

Note that these tests weren't actually testing the "only non-zero volume" logic previously (that logic is in the contracts), so I renamed them to be more clear about that.  I also removed a test that differed only by that, since it was a redundant test.