Renames variables to be more accurate.
I spent a lot of time trying to figure out what this code did because the variable name `_amountRemainingFillerWants` implies something different from what the variable actually contains.  In reality, this value is the min of "number of shares the order filler wants to trade" and "number of shares the order creator wants to trade".  The variable name previously implied that it was the amount that the filler wants only, which isn't true.  It is worth noting that the code `uint256 _amountFilled = _amountFillerWants.sub(_tradeSize);` is now more clearly incorrect, really that variable holds `_amountFillerWantsRemaining`, so it shouldn't be passed to `updateProfitLoss`!  I have fixed this as well.